refactor: remove SceneEventBus, fix scene API mismatches, add nametag system
Engine:
- Remove SceneEventBus over-abstraction; replace with Scene::m_PendingScenePath
- SceneTransitionSystem::Update returns std::optional<string> instead of publishing
- Fix Scene missing UpdateCommon declaration; inline into OnUpdateRuntime/OnUpdateSimulation
- Fix NetworkSystem calls to use GetInstance() (now member functions, not static)
- Remove PhysicsBodySystem::RegisterObservers (removed from API)
- Add audio_system.h include to scene.cpp
- Add SystemManager, SceneEventBus removal, registry facade to Scene
- Add Scene::LoadSceneAsync, SwapScene methods
- Refactor service_locator, service_registry
- Add asset_path_resolver
- Add text_renderer
- Add entity_collector, geometry_factory, lighting_manager, material_manager

Scene:
- Add nametag system with chshader materials
- Add scripting_components
- Refactor prefab_serializer
- Update scene_serializer, component_serializer
- Update physics_body_system, network_system

Editor:
- Remove SetEventCallback calls (EventBus removed)
- Add pending scene path polling in EditorLayer::OnUpdate
- Update viewport_panel, scene_manager, world_panel

Scripting:
- Add script_glue_animation, script_glue_physics, script_glue_player, script_glue_transform
- Refactor scriptengine, scene_scripting_manager
- Update managed C# components

Game:
- Update scenes, prefabs, scripts, materials

+++ engine/scene/component_ui_control.cpp
@@ -21,7 +21,7 @@ namespace Chained
 			std::vector<uint64_t> ids;
 			for (auto ent : s->GetRegistry().view<UIControlComponent>())
 			{
-				ids.push_back((uint64_t)(uint32_t)ent);
+				ids.push_back(static_cast<uint64_t>(ent));
 			}
 			return ids;
 		};
